Fix benchmark tests
