--- tests/test.rs
@@ -1,5 +1,0 @@
-extern crate racer;
-extern crate rand;
-
-#[cfg(test)] pub mod system;
-#[cfg(test)] pub mod bench;
